requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt[crypto]
django-filter
requests
daphne
django-heroku